Move get* functions to mem.c
