+++ contrib/fonttools/dewoff.c
@@ -4,6 +4,8 @@
 #include <stdlib.h>
 #include <zlib.h>
 
+#include "mem.h"
+
 /* Compile with: $ cc -o dewoff dewoff.c -lz */
 
 /* http://people.mozilla.com/~jkew/woff/woff-2009-09-16.html */
