+++ contrib/fonttools/findtable.c
@@ -1,6 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+
+#include "mem.h"
+
 typedef unsigned int uint32;
 typedef int int32;
 typedef short int16;
