+++ contrib/fonttools/stripttc.c
@@ -2,6 +2,8 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "mem.h"
+
 /* This program takes a ttc file and turns it into its component ttf files */
 /* The makes two changes to the data:                                      */
 /*	* The tables are placed a different offsets                        */
