+++ contrib/fonttools/ttf2eps.c
@@ -29,6 +29,8 @@
 #include <stdio.h>
 #include <time.h>
 
+#include "mem.h"
+
 typedef struct basepoint {
     double x, y;
 } BasePoint;
