+++ contrib/fonttools/woff.c
@@ -4,6 +4,8 @@
 #include <stdlib.h>
 #include <zlib.h>
 
+#include "mem.h"
+
 /* Compile with: $ cc -o woff woff.c -lz */
 
 /* http://people.mozilla.com/~jkew/woff/woff-2009-09-16.html */
