+++ fontforge/fvimportbdf.c
@@ -34,6 +34,7 @@
 #include "fontforgevw.h"
 #include "fvfonts.h"
 #include "macbinary.h"
+#include "mem.h"
 #include "namelist.h"
 #include "palmfonts.h"
 #include <gfile.h>
