+++ fontforge/ikarus.c
@@ -29,6 +29,7 @@
 
 #include "fontforge.h"
 #include "namelist.h"
+#include "mem.h"
 #include <utype.h>
 #include <string.h>
 #include <ustring.h>
