+++ fontforge/macbinary.c
@@ -35,6 +35,7 @@
 #include "fvfonts.h"
 #include "http.h"
 #include "lookups.h"
+#include "mem.h"
 #include "parsepfa.h"
 #include <stdlib.h>
 #include <string.h>
