+++ fontforge/splinesaveafm.c
@@ -31,6 +31,7 @@
 #include "fvfonts.h"
 #include "lookups.h"
 #include "macbinary.h"
+#include "mem.h"
 #include "mm.h"
 #include "namelist.h"
 #include "ttf.h"		/* For AnchorClassDecompose */
