+++ fontforge/ttfspecial.c
@@ -25,6 +25,7 @@
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #include "fontforge.h"
+#include "mem.h"
 #include <math.h>
 #include <time.h>
 #include <utype.h>
