+++ fontforge/woff.c
@@ -31,6 +31,7 @@
 
 #include "fontforge.h"
 #include "http.h"
+#include "mem.h"
 #include <math.h>
 #include <ctype.h>
